# all: Use `gomaxprocs` library to set GOMAXPROCS