--- a/beater/beater.go
+++ b/beater/beater.go
@@ -34,6 +34,7 @@ import (
 	"github.com/pkg/errors"
 	"go.elastic.co/apm/module/apmhttp/v2"
 	"go.elastic.co/apm/v2"
+	"go.uber.org/automaxprocs/maxprocs"
 	"golang.org/x/sync/errgroup"
 
 	"github.com/elastic/beats/v7/libbeat/beat"
@@ -95,11 +96,22 @@ func NewCreator(args CreatorParams) beat.Creator {
 			libbeatMonitoringRegistry: monitoring.Default.GetRegistry("libbeat"),
 		}
 
+		// Use `maxprocs` to change the GOMAXPROCS respecting any CFS quotas, if
+		// set. This is necessary since the Go runtime will default to the number
+		// of CPUs available in the  machine it's running in, however, when running
+		// in a container or in a cgroup with resource limits, the disparity can be
+		// extreme.
+		// Having a significantly greater GOMAXPROCS set than the granted CFS quota
+		// results in a significant amount of time spent "throttling", essentially
+		// pausing the the running OS threads for the throttled period.
+		_, err := maxprocs.Set(maxprocs.Logger(logger.Infof))
+		if err != nil {
+			logger.Errorf("failed to set GOMAXPROCS: %v", err)
+		}
 		var elasticsearchOutputConfig *agentconfig.C
 		if hasElasticsearchOutput(b) {
 			elasticsearchOutputConfig = b.Config.Output.Config()
 		}
-		var err error
 		bt.config, err = config.NewConfig(bt.rawConfig, elasticsearchOutputConfig)
 		if err != nil {
 			return nil, err