@simitt 2022-05-24
Given the issues that @marclop encountered with packaging local apm-servers, I suggest to mark this as ready for review and get it merged into `main` (`8.3`), and then define a test plan which deployment sizes to test on which cloud providers and instance sizes. The cloud benchnmarks should be run timely after FF to be able to react on any issues we might detect. 
@axw @marclop any objections on this plan?