@axw 2022-06-13
Removing test-plan, will be covered by https://github.com/elastic/apm-server/pull/8278